Apply Copilot + Gemini review feedback
- formatters.mts (resolve --fetch): apply same noBody dedup for first-look
  threads already shown in ## Review threads to resolve, matching fix-formatter
  (Copilot: consistent dedup behavior across both CLI commands)
- fix-formatter.mts: move resolutionOnlyIds Set creation inside
  firstLookThreads.length > 0 guard to avoid unnecessary work when only
  firstLookComments are present (Gemini: avoid unnecessary computation)
